+++ dokan-sys/CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Raw binding for `DokanRequestUnmount`.
+
 ### Changed
 
-- Upgrade to **Dokan 2.3.0**.
+- Upgrade to **Dokan 2.3.1**.
 - Bump dependencies.
 
 ## [0.3.1] - 2022-10-04